requirements.txt
requests>=2.31.0
GitPython>=3.1.40
PyYAML>=6.0.1
packaging>=23.0

packaging
